requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
flask>=3.0.0
apscheduler>=3.10.0
tzdata>=2024.1
